Remove pre-test.js
- New function to deploey ERC20 token contract
- Move token contract deployment to test.js
